+++ robo4j-hw-rpi/src/main/java/com/robo4j/hw/rpi/imu/bno/Bno055SelfTestResult.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Robo4J. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.robo4j.hw.rpi.imu;
+package com.robo4j.hw.rpi.imu.bno;
 
 /**
  * The result of a BNO055 self test.
