+++ robo4j-hw-rpi/src/main/java/com/robo4j/hw/rpi/imu/bno/impl/Bno055I2CDevice.java
@@ -14,13 +14,13 @@
  * You should have received a copy of the GNU General Public License
  * along with Robo4J. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.robo4j.hw.rpi.imu.impl;
+package com.robo4j.hw.rpi.imu.bno.impl;
 
 import com.pi4j.io.i2c.I2CBus;
 import com.pi4j.io.i2c.I2CDevice;
 import com.pi4j.io.i2c.I2CFactory;
 import com.robo4j.hw.rpi.i2c.ReadableDevice;
-import com.robo4j.hw.rpi.imu.Bno055Device;
+import com.robo4j.hw.rpi.imu.bno.Bno055Device;
 import com.robo4j.math.geometry.Tuple3f;
 
 import java.io.IOException;
