+++ robo4j-hw-rpi/src/main/java/com/robo4j/hw/rpi/imu/bno/shtp/ShtpPacketRequest.java
@@ -17,7 +17,7 @@
 
 package com.robo4j.hw.rpi.imu.bno.shtp;
 
-import static com.robo4j.hw.rpi.imu.impl.Bno080SPIDevice.SHTP_HEADER_SIZE;
+import static com.robo4j.hw.rpi.imu.bno.impl.Bno080SPIDevice.SHTP_HEADER_SIZE;
 
 /**
  * ShtpPacketRequest represent SHTP request format used by SPI devices
